fritzcreator: replace obscure padding generation with something more portable
It was reported that OpenWrt/Gluon would sometimes generate invalid
bootloader images with insufficient padding for the FritzBox [1],
presumably depending on the distro version used on the build host.

Fix this by replacing the printf-based padding generation (which in one
case relied on 65536 arguments being passed to a printf command) with a
more portable read from /dev/zero.

[1] https://github.com/freifunk-gluon/gluon/issues/1766